requirements.txt.txt
Flask
requests
python-dotenv
google-generativeai
flask-cors
